fix(backend): fix image service syntax errors and add asset CRUD handlers
- Fix broken syntax in image_service.go with proper imaging library usage
- Fix ImageDimensions.Size type to int64 for consistency
- Add missing asset CRUD handler methods (ListAssets, GetAsset, CreateAsset, UpdateAsset, DeleteAsset)
- Add github.com/disintegration/imaging dependency for image processing

+++ internal/domain/image_asset_models.go
@@ -13,9 +13,9 @@ const (
 
 // ImageDimensions 图片尺寸信息
 type ImageDimensions struct {
-	Width  int `json:"width"`
-	Height int `json:"height"`
-	Size   int `json:"size"` // 字节大小
+	Width  int   `json:"width"`
+	Height int   `json:"height"`
+	Size   int64 `json:"size"` // 字节大小
 }
 
 // ImageInfo 图片信息
